not sure
did a lot of stuff to try and figure out what's going on. but the basic
lantz query function they suggest in the driver tutorial makes it
timeout and not communicate with the instrument.

still want to improve the design to follow proper coding practice better
but not playing nice with whatever the @Feats() and __init__ functions
do. basically confused by lanz