added sensitivities to SR7265, SP2150 is still derp
this driver should be done...still missing comments and some logging
functionality, but it should be able to be useful for most stuff you
might want to do experimentally.

SP2150 I can communicate w/ over USB, but there are some weird things
going on where it drops commands or something. still debugging this.